Added new abstract class for column rendering
